+++ inst/tinytest/count/count_number_of_tests.R
@@ -64,6 +64,6 @@ print(testcount_total)
 
 # test / function ratio ====
 nfuns <- length(getNamespaceExports(loadNamespace("tinycodet")))
-testcount_total / nfuns # about 48 tests per function on average
+testcount_total / nfuns # about 78 tests per function on average