+++ inst/tinytest/regular/test-decimal-eq.R
@@ -80,3 +80,4 @@ expect_error(1 %d<% "a")
 expect_error(1 %d>% "a")
 expect_error(1 %d{}% "a")
 expect_error(1 %d!{}% "a")
+